Including David M. Gay's dtoa.c to get forward with FP support. See full commit message for notes.
Notes:

dtoa.c is NOT public domain / CC0. It comes with its own license -- a
modified MIT license, as far as I can tell. Similar to how I included
Doug Lea's malloc(), I added a patch file to make it easier to reapply
those changes to newer versions of Mr. Gay's work.

dtoa.c includes the standard function strtod(), and dtoa(), which can
be utilized to cover FP output in the *printf() function family. With
this commit, only strtod() is added; integration of that into *scanf()
and of dtoa() into *printf() is the next TODO on the list.